Refactor: moved Registration logic to aiosip
- Created a new file registration.py in the aiosip package with Registration logic.
- Updated __init__.py to include registration module.
- Refactored examples to import Registration from aiosip.registration.
- Removed util.py as it's no longer needed.
